containers/flask-container/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
click==8.1.7
disutils==1.4.32.post2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-PyMongo==2.3.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.0
multidict==6.1.0
numpy==2.1.2
pandas==2.2.3
propcache==0.2.0
py4j==0.10.9.7
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
yarl==1.14.0
python-dotenv==1.0.1
pymongo==3.12.0
marshmallow==3.22.0